Split out light state - share light state between light node and scenes
